mr_browse_test: fix test code
The changes from commit 9fa369859 changed the expected behavior on
Test_mrBrowseCurrent test code, which was waiting for an MR created from
the upstream repository instead of the fork that the 'mrtest' branch
tracks. This patch fix it by checking-out to the 'mrtest2' branch and
waiting for MR 3, which was created from fork's 'mrtest2' branch.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>