config: Add config verification
From GitLab:

"The ability to create personal access tokens without expiry was
deprecated in GitLab 15.4 and removed in GitLab 16.0. In GitLab 16.0 and
later, existing personal access tokens without an expiry date are
automatically given an expiry date of 365 days later than the current
date. The automatic adding of an expiry date occurs on GitLab.com during
the 16.0 milestone. The automatic adding of an expiry date occurs on
self-managed instances when they are upgraded to GitLab 16.0. This change
is a breaking change."

Frequent users of GitLab are being caught by this 1-year expiration, and
are seeing general command errors, for example,

[prarit@prarit centos-stream-9]$ lab mr show 123
2023/05/20 08:34:51 ERROR: mr_show.go:45: GET https://gitlab.com/api/v4/projects/redhat/centos-stream/src/kernel/centos-stream-9/merge_requests/123: 401 {message: 401 Unauthorized}

Do a once-a-day check to see if a user's token is valid by querying for
the user's userID directly from the GitLab instance.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>