lab: General fixes for Labels
After the package update the compile fails with many warnings due to a
change in the way Labels are handled in the GitLab API.  There are now two
methods used, one which uses a []gitlab.Label array and the other which
uses a type Labels []string array.

This change means we now have to carry two different functions for
fetching label data, mapLabelsAsStrings() and mapLabelsAsLabels().  While the
function names are wonky, they do represent what they do in the code.

Move and rename the existing mapLabels to utils.go:mapLabelsAsStrings and add
a utils.go:mapLabelsAsLabels() function.  Update the labels code to use
the new functions.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>