internal/config/config.go: Warn on expiring tokens
82444bfa9c07 ("config: Add config verification") added some simple
checking to see if a user's token was valid.  After the recent go module
updates we can now use the PersonalAccessToken service from go-gitlab to
warn a user that their token is about to expire.  This warning should
happen once a day when the config is verified.

Warn a user when their token is going to expire within two weeks.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>