lab: Add show mr approval rules
The WebUI provides a mechanism to view the status of MR approval rules on a
specific MR.  Add this functionality to lab.

Usage: 'lab mr approval-rule 1234'

One thing to note is that, unlike the WebUI, lab currently does not show
the 'Commented By' data.  This is because getting the data is expensive,
and at this time provides little functionality.  This can be added if
users request it.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>