gitlab: Fix 'mr list' limits
The command 'lab mr list --author $author_name --state merged' was
returning a list of 100 for each engineer on a project.  Evidence in the
git tree showed that this was not the case (most had more than 100).

This was due to a calcluational error in the MRList() code.  Implement a
new algorithm that correctly returns all MRs instead of the last 100.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>