cmd/mr_discussion: populate PositionOptions when needed
With commit 4d1dbfa6befe ("cmd/mr_discussion.go: Switch from NotePosition to
PositionOptions") the PositionOptions object was being created
regardless the presence of --position in user's command, causing the
REST API to fail due to the presence of a "position" parameter in the
request without all required arguments. This commit creates the
PositionOptions object within the --position handler, ensuring it's not
passed "empty".

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>