ci_view_test: minimal change TUI chars
Due to a minimal change in tview library an additional char was added
when printing the stage and jobs to the terminal. Unfortunately these
tests are too sensitive to this type of changes.

My suggestion is that we remove visual tests from our code, since the
library has their own tests and we should not be worried on how things
are really printed.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>