cmd/ci_status.go: Add color to ci status output
'lab ci status' shows status for each job as well as each pipeline.  There
are various states (success, failed, skipped, etc.) that can be output.

Currently the output is in white.  It would be nice if these were easier
to read in color: red for failed, green for success, and yellow for the
various 'in progress' states.

Output the ci status in color.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>