cmd/mr_show.go: Add CI Status to the 'mr show' output
If I use the webUI, I can see a 'green checkmark' or a 'red x' that
indicates the CI status.  That is not available with 'lab mr show'.

It's worthwhile to make this call, even though this takes another gitlab
call.  The zero entry in the jobs is the main job for the MR.  All other
jobs point to that job, so it shows the overall status for the MR.

Add CI Status to the output.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>