cmd/mr_show.go: Colorize Status output
The webUI shows Status, Merge Status, and CI Status and green, yellow, or
red.  This is a nice feature as it visually alerts users to what the
overall status of a Merge Request is.

Colorize the Status, Merge Status, and CI Status output.

Additional fix: While changing this code I noticed that
DetailedMergeStatus has many more values than it did before [1].  The only
one of these that should be 'green' is mergeable.  The other statuses
should all be red.

[1] https://docs.gitlab.com/api/merge_requests/#single-merge-request-response-notes

Signed-off-by: Prarit Bhargava <prarit@redhat.com>