cmd/mr_list.go: Add 'fancy' output
I've always found the project merge_requests page useful.  It shows the
status of MRs, the CI and Approval statuses, etc.  Let's start doing that
in the 'mr list' command.  I still like the default simple command, so
let's add a '--show-state' to get fancy output.

Note, that this requires a per-MR lookup to get the CI status.  That
slows the output of --show-state down, but it's worth the trade off.

Add --show-state to output similar to the project merge_requests page.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>