cmd/mr_list: Fix mr list output
During the work on --show-status, it was noticed that gitlab calls were
taking a long time.  This was because the 'number' parameter was not being
properly dealt with in the code.  One key issue was an == comparison vs
a >= comparison in MRList().

It was also noticed that the --all feature really serves no purpose.  The
default is supposed to be to show all MRs, so the --all option can be
deprecated.

Fix mr list output.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>