cmd/mr_list.go: Add Author to 'mr list --show-status' output
Add the MR Author's name to the title field.  I thought about adding
yet-another-column, but it looked to busy.  So instead, add the MR
Author's name in brackets beside the title.  This does have the
side-effect of shortening the title but I think it is a good trade-off vs
another column.

Add Author to 'mr list --show-status' pretty output.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>